Move BaseUpdater in a separate file
